move versioning to top level package
