+++ scripts/build.sh
@@ -1,5 +1,5 @@
-#!/bin/sh
+#!/bin/bash
 
 set -eo pipefail
 
-go build -ldflags "-X github.com/bryanl/doit/commands.Build=`git rev-parse HEAD`" github.com/bryanl/doit/cmd/doit
+go build -ldflags "-X github.com/bryanl/doit.Build=`git rev-parse HEAD`" github.com/bryanl/doit/cmd/doit
